dom/imptests/html/html/dom/elements/global-attributes/test_document-dir.html

<!DOCTYPE html>
<html dir="LTR">
<title>document.dir</title>
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-dir">
<link rel="help" href="http://www.whatwg.org/html5/#reflect">
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="stylesheet" href="/resources/testharness.css">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_equals(document.dir, "ltr");
  assert_equals(document.documentElement.getAttribute("dir"), "LTR");
}, "Markup attribute")
test(function() {
  document.dir = "x-garbage";
  assert_equals(document.dir, "");
  assert_equals(document.documentElement.getAttribute("dir"), "x-garbage");
}, "Setting the idl attribute to a garbage value")
test(function() {
  document.dir = "";
  assert_equals(document.documentElement.hasAttribute("dir"), false, "Attribute should be gone");
}, "Setting the idl attribute to the empty sting")
</script>
